Renamed namespace of controllers to avoid conflict between namespace and the admin entity class.
