+++ FribergCarRentalsBravo/Controllers/AdminController .cs
@@ -9,7 +9,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 
-namespace FribergCarRentalsBravo.Controllers.Admin
+namespace FribergCarRentalsBravo.Controllers.Admins
 {
     public class AdminController : Controller
     {
