+++ FribergCarRentalsBravo/Controllers/Admins/AdminCustomerController.cs
@@ -1,39 +1,33 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using FribergCarRentalsBravo.DataAccess.DatabaseContexts;
+﻿using FribergCarRentalsBravo.Data;
 using FribergCarRentalsBravo.DataAccess.Entities;
 using FribergCarRentalsBravo.DataAccess.Repositories;
-using FribergCarRentalsBravo.Data;
 using FribergCarRentalsBravo.Helpers;
 using FribergCarRentalsBravo.Sessions;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 
-namespace FribergCarRentalsBravo.Controllers.Admin
+namespace FribergCarRentalsBravo.Controllers.Admins
 {
-    public class AdminCarCategoryController : Controller
+    public class AdminCustomerController : Controller
     {
         #region Fields
 
-        private readonly ICarCategoryRepository carCategoryRepo;
+        public ICustomerRepository customerRep { get; }
 
         #endregion
 
-        #region Constructors        
+        #region Constructors
 
-        public AdminCarCategoryController(ICarCategoryRepository carCategoryRepo)
+        public AdminCustomerController(ICustomerRepository customerRep)
         {
-            this.carCategoryRepo = carCategoryRepo;
+            this.customerRep = customerRep;
         }
 
         #endregion
 
         #region Actions
 
-        // GET: CarCategory
+        // GET: CustomerController
         public async Task<IActionResult> Index()
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
@@ -41,10 +35,11 @@ namespace FribergCarRentalsBravo.Controllers.Admin
                 return RedirectToLogin(nameof(Index));
             }
 
-            return View(await carCategoryRepo.GetAllAsync());
+            var customers = await customerRep.GetAllCustomers();
+            return View(customers);
         }
 
-        // GET: CarCategory/Details/5
+        // GET: CustomerController/Details/5
         public async Task<IActionResult> Details(int id)
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
@@ -52,23 +47,11 @@ namespace FribergCarRentalsBravo.Controllers.Admin
                 return RedirectToLogin(nameof(Details));
             }
 
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var carCategory = await carCategoryRepo.GetByIdAsync(id);
-
-            if (carCategory == null)
-            {
-                return NotFound();
-            }
-
-            return View(carCategory);
+            return View(await customerRep.GetCustomerById(id));
         }
 
-        // GET: CarCategory/Create
-        public async Task<IActionResult> Create()
+        // GET: CustomerController/Create
+        public IActionResult Create()
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
             {
@@ -78,12 +61,10 @@ namespace FribergCarRentalsBravo.Controllers.Admin
             return View();
         }
 
-        // POST: CarCategory/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: CustomerController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("CarCategoryId,Name")] CarCategory carCategory)
+        public async Task<IActionResult> Create(Customer customer)
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
             {
@@ -92,14 +73,16 @@ namespace FribergCarRentalsBravo.Controllers.Admin
 
             if (ModelState.IsValid)
             {
-                await carCategoryRepo.CreateNewCarCategoryAsync(carCategory);
+                await customerRep.CreateCustomer(customer);
                 return RedirectToAction(nameof(Index));
             }
-
-            return View(carCategory);
+            else
+            {
+                return View();
+            }
         }
 
-        // GET: CarCategory/Edit/5
+        // GET: CustomerController/Edit/5
         public async Task<IActionResult> Edit(int id)
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
@@ -107,45 +90,51 @@ namespace FribergCarRentalsBravo.Controllers.Admin
                 return RedirectToLogin(nameof(Edit));
             }
 
-            if (id == null)
+            if (id == null || customerRep.GetAllCustomers == null)
             {
                 return NotFound();
             }
 
-            var carCategory = await carCategoryRepo.GetByIdAsync(id);
-            if (carCategory == null)
+            Customer customer = await customerRep.GetCustomerById(id);
+
+            if (customer == null)
             {
                 return NotFound();
             }
-            return View(carCategory);
+            return View(customer);
         }
 
-        // POST: CarCategory/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: CustomerController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("CarCategoryId,Name")] CarCategory carCategory)
+        public async Task<IActionResult> Edit(int id, Customer customer)
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
             {
                 return RedirectToLogin(nameof(Edit));
             }
 
-            if (id != carCategory.CarCategoryId)
+            if (id != customer.CustomerId)
             {
                 return NotFound();
             }
 
             if (ModelState.IsValid)
             {
-                carCategoryRepo.UpdateCarCategoryrAsync(carCategory);
+                try
+                {
+                    await customerRep.EditCustomer(customer);
+                }
+                catch (Exception)
+                {
+                    return View();
+                }
                 return RedirectToAction(nameof(Index));
             }
-            return View(carCategory);
+            return View();
         }
 
-        // GET: CarCategory/Delete/5
+        // GET: CustomerController/Delete/5
         public async Task<IActionResult> Delete(int id)
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
@@ -153,37 +142,41 @@ namespace FribergCarRentalsBravo.Controllers.Admin
                 return RedirectToLogin(nameof(Delete));
             }
 
-            if (id == null)
+            Customer customer = await customerRep.GetCustomerById(id);
+            if (customer == null)
             {
                 return NotFound();
             }
-
-            var carCategory = await carCategoryRepo.GetByIdAsync(id);
-            if (carCategory == null)
-            {
-                return NotFound();
-            }
-
-            return View(carCategory);
+            return View(customer);
         }
 
-        // POST: CarCategory/Delete/5
+        // POST: CustomerController/Delete/5             
+
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             if (!UserSessionHandler.IsAdminLoggedIn(HttpContext.Session))
             {
-                return RedirectToLogin(nameof(DeleteConfirmed));
+                return RedirectToLogin(nameof(Delete));
             }
 
-            var carCategory = await carCategoryRepo.GetByIdAsync(id);
-            if (carCategory != null)
+            var customer = await customerRep.GetCustomerById(id);
+            if (customer == null)
             {
-                await carCategoryRepo.DeleteCarCategoryAsync(carCategory);
+                return NotFound();
             }
 
-            return RedirectToAction(nameof(Index));
+            try
+            {
+                await customerRep.DeleteCustomer(customer);
+                return RedirectToAction(nameof(Index));
+            }
+            catch (Exception)
+            {
+                ModelState.AddModelError(string.Empty, "Det gick inte att ta bort kunden. Försök igen senare.");
+                return View(customer);
+            }
         }
 
         #endregion
@@ -194,14 +187,14 @@ namespace FribergCarRentalsBravo.Controllers.Admin
         /// Redirects to the login page and request a redirect back afterwards. 
         /// </summary>
         /// <param name="action">The action to redirect to.</param>
-        /// <param name="id">An optional ID for the category.</param>
+        /// <param name="id">An optional ID for the customer.</param>
         /// <returns><see cref="IActionResult"/>.</returns>
         private IActionResult RedirectToLogin(string action, int? id = null)
         {
             RouteValueDictionary? routeValues = id is not null ? new RouteValueDictionary(new { id = id }) : null;
 
             TempDataHelper.Set(TempData, AdminController.RedirectToPageTempDataKey, new RedirectToActionData(
-                    action, ControllerHelper.GetControllerName<AdminCarCategoryController>(), routeValues: routeValues));
+                    action, ControllerHelper.GetControllerName<AdminCustomerController>(), routeValues: routeValues));
 
             return RedirectToAction(nameof(AdminController.Login), ControllerHelper.GetControllerName<AdminController>());
         }
